+++ inst/shiny/modules/tab_nca.R
@@ -74,7 +74,7 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
     adnca_data <- reactive(pknca_data()$conc$data)
 
     # #' NCA Setup module
-    nca_setup <- setup_server("nca_setup", adnca_data, pknca_data)
+    nca_setup <- setup_server("nca_setup", adnca_data, pknca_data, extra_group_vars)
 
     processed_pknca_data <- nca_setup$processed_pknca_data
     settings <- nca_setup$settings
@@ -138,7 +138,9 @@ tab_nca_server <- function(id, pknca_data, extra_group_vars) {
               check_reasons = TRUE
             ) %>%
             # Perform PKNCA parameter calculations
-            PKNCA_calculate_nca() %>%
+            PKNCA_calculate_nca(
+              blq_rule = settings()$data_imputation$blq_imputation_rule
+            ) %>%
             # Add bioavailability results if requested
             add_f_to_pknca_results(settings()$bioavailability) %>%
             # Apply standard CDISC names